% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_net.R
\name{df_to_mat}
\alias{df_to_mat}
\title{Converts data.frames to matrices amenable to network analysis with btergm}
\usage{
df_to_mat(unit_time, dyad_time)
}
\arguments{
\item{unit_time}{data.frame unit/time dataset with columns named `unit`,
`time`. Additional columns are vertex attributes}

\item{dyad_time}{data.frame dyadic dataset with columns named `unit1`,
`unit2`, `time`. Additional columns are used to create the endogenous and
exogenous networks}
}
\description{
Converts data.frames to matrices amenable to network analysis with btergm
}
\examples{
# Examples are in the README file at
http://github.com/vincentarelbundock/btergmHelper
}
