% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_net.R
\name{df_to_net}
\alias{df_to_net}
\title{Align and subset matrices using a vector of unit names}
\usage{
df_to_net(endo_net, unit_time, dyad_time, ...)
}
\arguments{
\item{endo_net}{character name of the dependent variable (from `dyad_time`)}

\item{unit_time}{data.frame unit/time dataset with columns named `unit`,
`time`. Additional columns are vertex attributes}

\item{dyad_time}{data.frame dyadic dataset with columns named `unit1`,
`unit2`, `time`. Additional columns are used to create the endogenous and
exogenous networks}
}
\description{
Align and subset matrices using a vector of unit names
}
\note{
unnamed arguments (e.g., `directed`, `loops`) will be passed to the
`network::network` function
}
\examples{
# Examples are in the README file at
http://github.com/vincentarelbundock/btergmHelper
}
