% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{dyadic_w}
\alias{dyadic_w}
\title{Create a new Wy column to measure specific/aggregate source/target contagion
(dyadic cross-sectional data)}
\usage{
dyadic_w(dat, source = "unit1", target = "unit2", y = "y", w = "w",
  wy = "wy", type = "specific_source", weights = "ik",
  row_normalize = TRUE, zero_loop = TRUE, ncpus = 1,
  progress = TRUE)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{source}{name of source id column (character)}

\item{target}{name of target id column (character)}

\item{y}{name of outcome to lag (character)}

\item{w}{name of distance/weights column (character)}

\item{wy}{name of the output variable (character)}

\item{type}{(character) aggregate source (different source); aggregate
target (different target); specific source (same target different source);
specific target (same source different target)}

\item{weights}{weight specification: ik, jk, im, jm (character)}

\item{row_normalize}{should the W matrix be row-normalized? (boolean)}

\item{zero_loop}{should wy be set to 0 when source == target (boolean)}

\item{ncpus}{number of cpus to use for parallel computation (integer)}

\item{progress}{show progress bar (boolean)}
}
\description{
Create a new Wy column to measure specific/aggregate source/target contagion
(dyadic cross-sectional data)
}
