% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{dyadic_wy}
\alias{dyadic_wy}
\title{Adds a new Wy column which measures specific/aggregate source/target contagion
(directed dyadic cross-sectional or panel data)}
\usage{
dyadic_wy(dat, source = "unit1", target = "unit2", y = "y",
  w = "w", wy = "wy", time = NULL, type = "specific_source",
  weights = "ik", row_normalize = TRUE, zero_loop = TRUE,
  ncpus = 1, progress = TRUE)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{source}{name of source id column (character)}

\item{target}{name of target id column (character)}

\item{y}{name of outcome to lag (character)}

\item{w}{name of distance/weights column (character)}

\item{wy}{name of the output variable (character)}

\item{time}{name of the time variable (optional) (character)}

\item{type}{(character) aggregate source (different source); aggregate
target (different target); specific source (same target different source);
specific target (same source different target)}

\item{weights}{weight specification: ik, jk, im, jm (character)}

\item{row_normalize}{should each value of the W matrix be divided by the
row-wise sum? (boolean)}

\item{zero_loop}{should wy be set to 0 when source == target (boolean)}

\item{ncpus}{number of cpus to use for parallel computation (integer)}

\item{progress}{show progress bar (boolean) (only works for cross-sectional
data.)}
}
\description{
Adds a new Wy column which measures specific/aggregate source/target contagion
(directed dyadic cross-sectional or panel data)
}
