% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_net.R
\name{mat_to_net}
\alias{mat_to_net}
\title{Converts a list of matrices and vertex attributes to a dependent network for
btergm}
\usage{
mat_to_net(dv, iv, ...)
}
\arguments{
\item{dv}{list of matrix stored in the environment produced by `df_to_mat`}

\item{iv}{list of vertex attributes stored in the `vertex.attributes` object
in the environment produced by `df_to_mat`}
}
\description{
Converts a list of matrices and vertex attributes to a dependent network for
btergm
}
\note{
unnamed arguments (e.g., `directed`, `loops`) will be passed to the
`network::network` function that is used under the hood to create network
objects.
}
