% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_to_network.R
\name{panel_to_network}
\alias{panel_to_network}
\title{Converts data.frames to objects amenable to network analysis using `btergm`}
\usage{
panel_to_network(unit_time, dyad_time, cores = 1, verbose = TRUE)
}
\arguments{
\item{unit_time}{data.frame unit/time dataset with columns named `unit`,
`time`. Additional columns are vertex attributes}

\item{dyad_time}{data.frame dyadic dataset with columns named `unit1`,
`unit2`, `time`. Additional columns are edge attributes.}

\item{cores}{integer number of cores to use for computation with mclapply}

\item{verbose}{print progress report if TRUE}
}
\description{
Converts data.frames to objects amenable to network analysis using `btergm`
}
\examples{
# Examples are in the README file at
http://github.com/vincentarelbundock/btergmHelper
}
