% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{sanity}
\alias{sanity}
\title{Internal sanity check function}
\usage{
sanity(dat, source = "source", target = "target", w = "w",
  y = NULL, time = NULL)
}
\description{
Internal sanity check function
}
