% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{sl_dyad_specific_panel}
\alias{sl_dyad_specific_panel}
\title{Produce a new Wy column to measure specific source/target contagion (panel
data)}
\usage{
sl_dyad_specific_panel(dat, unit_src, unit_tar, time, w, y, type, zero_diag,
  row_normalize)
}
\arguments{
\item{dat}{dyad-year dataset (data.frame)}

\item{unit_src}{name of source id column (character)}

\item{time}{name of time id column (character)}

\item{w}{name of distance/weights column (character)}

\item{y}{name of outcome to lag (character)}

\item{type}{"source" or "target"-specific contagion (character)}

\item{zero_diag}{should the W matrix diagonal be zeroed? (boolean)}

\item{row_normalize}{should the W matrix be row-normalized? (boolean)}

\item{unit_target}{name of target id column (character)}
}
\description{
Produce a new Wy column to measure specific source/target contagion (panel
data)
}
