% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{specific_source}
\alias{specific_source}
\title{Add a new Wy column to measure specific source/target contagion
(cross-sectional data)}
\usage{
specific_source(dat, unit_src = "unit1", unit_tar = "unit2", w = "w",
  y = "y", wy = "wy", row_normalize = TRUE, full = FALSE)
}
\arguments{
\item{dat}{dyad-year dataset (data.frame)}

\item{unit_src}{name of source id column (character)}

\item{w}{name of distance/weights column (character)}

\item{y}{name of outcome to lag (character)}

\item{wy}{name of the output variable (character)}

\item{row_normalize}{should the W matrix be row-normalized? (boolean)}

\item{full}{if FALSE, only returns columns unit_src, unit_tar, time, wy (boolean)}

\item{unit_target}{name of target id column (character)}
}
\description{
Add a new Wy column to measure specific source/target contagion
(cross-sectional data)
}
