% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{specific_source_panel}
\alias{specific_source_panel}
\title{A wrapper around `specific_source`. Split-Apply-Combine by time period.}
\usage{
specific_source_panel(dat, unit_src = "unit1", unit_tar = "unit2",
  time = "time", w = "w", y = "y", row_normalize = TRUE)
}
\arguments{
\item{dat}{dyad-year dataset (data.frame)}

\item{unit_src}{name of source id column (character)}

\item{time}{name of time column (character)}

\item{w}{name of distance/weights column (character)}

\item{y}{name of outcome to lag (character)}

\item{row_normalize}{should the W matrix be row-normalized? (boolean)}

\item{unit_target}{name of target id column (character)}
}
\description{
A wrapper around `specific_source`. Split-Apply-Combine by time period.
}
