% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{undirected_to_directed}
\alias{undirected_to_directed}
\title{duplicating rows while inverting source and target identifiers.}
\usage{
undirected_to_directed(dat, source = "unit1", target = "unit2",
  time = NULL)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{source}{name of source id column (character)}

\item{target}{name of target id column (character)}
}
\description{
duplicating rows while inverting source and target identifiers.
}
