% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_to_network.R
\name{dependent_network}
\alias{dependent_network}
\title{Converts a list of matrices and vertex attributes to a dependent network for
btergm}
\usage{
dependent_network(network_dependent, network_environment, ...)
}
\arguments{
\item{network_dependent}{name of the dependent network (character)}

\item{network_environment}{an environment produced by the panel_to_network
function.}

\item{...}{arguments will be passed to the network::network function.}
}
\description{
Converts a list of matrices and vertex attributes to a dependent network for
btergm
}
\note{
unnamed arguments (e.g., `directed`, `loops`) will be passed to the
`network::network` function that is used under the hood to create network
objects.
}
