% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{dyadic_wy}
\alias{dyadic_wy}
\title{Adds a new Wy column which measures specific/aggregate source/target contagion
(directed dyadic cross-sectional or panel data)}
\usage{
dyadic_wy(dat, source = "unit1", target = "unit2", y = "y",
  w = "w", wy = "wy", time = NULL, type = "specific_source",
  weights = "ik", row_normalize = TRUE, zero_loop = TRUE,
  ncpus = 1, progress = TRUE, return_data = FALSE)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{source}{name of source id column (character)}

\item{target}{name of target id column (character)}

\item{y}{name of outcome to lag (character)}

\item{w}{name of distance/weights column (character)}

\item{wy}{name of the output variable (character)}

\item{time}{name of the time variable (optional) (character)}

\item{type}{of W matrix (character) 
\itemize{
\item specific source: y_ij = f(sum_k!=i w * y_kj)
\item specific target: y_ij = f(sum_m!=j w * y_im)
\item aggregate source: y_ij = f(sum_k!=i sum_m w * y_km)
\item aggregate target: y_ij = f(sum_k sum_m!=j w * y_km)
}}

\item{weights}{weight specification: ik, jk, im, jm (character)}

\item{row_normalize}{should each value of the W matrix be divided by the
row-wise sum? (boolean)}

\item{zero_loop}{should wy be set to 0 when source == target (boolean)}

\item{ncpus}{number of cpus to use for parallel computation (integer)}

\item{progress}{show progress bar (boolean) (only works for cross-sectional
data.)}

\item{return_data}{function returns the full dataset (TRUE) or just the
spatial lag (FALSE)}
}
\description{
Adds a new Wy column which measures specific/aggregate source/target contagion
(directed dyadic cross-sectional or panel data)
}
