% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{dyadic_wy}
\alias{dyadic_wy}
\title{Adds a new Wy column which measures specific/aggregate origin/destination contagion
(directed dyadic cross-sectional or panel data)}
\usage{
dyadic_wy(
  dat,
  origin = "unit1",
  destination = "unit2",
  y = "y",
  w = "w",
  wy = "wy",
  time = NULL,
  type = "specific_origin",
  weights = "ik",
  row_normalize = TRUE,
  zero_loop = TRUE,
  progress = FALSE
)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{origin}{name of origin id column (character)}

\item{destination}{name of destination id column (character)}

\item{y}{name of outcome to lag (character)}

\item{w}{name of distance/weights column (character)}

\item{wy}{name of the output variable (character)}

\item{time}{name of the time variable (optional) (character)}

\item{type}{of W matrix (character)
\itemize{
  \item specific origin: y_ij = f(sum_k!=i w * y_kj)
  \item specific destination: y_ij = f(sum_m!=j w * y_im)
  \item aggregate origin: y_ij = f(sum_k!=i sum_m w * y_km)
  \item aggregate destination: y_ij = f(sum_k sum_m!=j w * y_km)
}}

\item{weights}{weight specification: ik, jk, im, jm (character)}

\item{row_normalize}{should each value of the W matrix be divided by the
row-wise sum? (boolean)}

\item{zero_loop}{should wy be set to 0 when origin == destination (boolean)}

\item{progress}{boolean show progress bar (boolean). Only works with
`plan(multiprocess)` and panel data (`time != NULL`).}
}
\description{
Adds a new Wy column which measures specific/aggregate origin/destination contagion
(directed dyadic cross-sectional or panel data)
}
