% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{monadic_w}
\alias{monadic_w}
\title{Create a W matrix. With panel data, this will produce a block diagonal matrix.}
\usage{
monadic_w(dat, source = "unit1", target = "unit2", w = "w",
  time = NULL, row_normalize = TRUE)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{source}{name of source id column (character)}

\item{target}{name of target id column (character)}

\item{w}{name of distance/weights column (character)}

\item{time}{name of the (optional) time variable column (character)}

\item{row_normalize}{should each value of the W matrix be divided by the
row-wise sum? (boolean)}
}
\description{
Create a W matrix. With panel data, this will produce a block diagonal matrix.
}
