% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{monadic_w_cs}
\alias{monadic_w_cs}
\title{Internal function to apply to a cross-section}
\usage{
monadic_w_cs(dat, source = "unit1", target = "unit2", w = "w")
}
\description{
Internal function to apply to a cross-section
}
