% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_to_network.R
\name{panel_to_network}
\alias{panel_to_network}
\title{Converts data.frames to matrices amenable to network analysis with btergm}
\usage{
panel_to_network(unit_time, dyad_time, cores = 1)
}
\arguments{
\item{unit_time}{data.frame unit/time dataset with columns named `unit`,
`time`. Additional columns are vertex attributes}

\item{dyad_time}{data.frame dyadic dataset with columns named `unit1`,
`unit2`, `time`. Additional columns are edge attributes.}

\item{cores}{integer number of cores to use for computation with mclapply}
}
\description{
Converts data.frames to matrices amenable to network analysis with btergm
}
\examples{
# Examples are in the README file at
http://github.com/vincentarelbundock/btergmHelper
}
