% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lag.R
\name{sanity}
\alias{sanity}
\title{Internal sanity check function}
\usage{
sanity(
  dat,
  origin = NULL,
  destination = NULL,
  w = NULL,
  y = NULL,
  wy = NULL,
  time = NULL,
  weights = NULL,
  type = NULL
)
}
\arguments{
\item{dat}{directed dyadic dataset (data.frame)}

\item{origin}{name of origin id column (character)}

\item{destination}{name of destination id column (character)}

\item{w}{name of distance/weights column (character)}

\item{y}{name of outcome to lag (character)}

\item{wy}{name of the output variable (character)}

\item{time}{name of the time variable (optional) (character)}

\item{weights}{weight specification: ik, jk, im, jm (character)}

\item{type}{of W matrix (character) 
\itemize{
  \item specific origin: y_ij = f(sum_k!=i w * y_kj)
  \item specific destination: y_ij = f(sum_m!=j w * y_im)
  \item aggregate origin: y_ij = f(sum_k!=i sum_m w * y_km)
  \item aggregate destination: y_ij = f(sum_k sum_m!=j w * y_km)
}}
}
\description{
Internal sanity check function
}
